requirements.txt
pybind11
transformers
py-libnuma
git+https://github.com/EleutherAI/lm-evaluation-harness